scraper/requirements_analysis.txt
requests>=2.28.1
beautifulsoup4>=4.11.1
lxml>=4.9.1
tinydb>=4.7.0
python-dotenv>=0.21.0
free-proxy>=1.0.2
ujson>=5.5.0
retry>=0.9.2
argparse>=1.4.0
colorlog>=6.7.0
scraperapi-sdk>=0.2.2

requests-ip-rotator>=1.0.14
flask>=2.2.3
pandas>=1.5.3
matplotlib>=3.7.1
plotly>=5.14.0
dash>=2.9.1
numpy>=1.24.2
wordcloud>=1.8.2.2
